+++ HISTORY.md
@@ -2,10 +2,12 @@ History
 ===============
 
 ### By now 2014-03-14
-* 3 major enhancement
+* 2 major enhancement
 	* Completed the design of architectures
-	* Added BioTCM::Modules::Utility
-	* Added BioTCM::Modules::WorkingDir
+	* Completed following modules and classes
+		* Added BioTCM::Modules::Utility
+		* Added BioTCM::Modules::WorkingDir
+		* Added BioTCM::Table
 
 ### 0.0.0 2014-03-06
 * 1 major enhancement